lab_4/app/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.1
marshmallow==3.21.2
PyMySQL==1.1.1
SQLAlchemy==2.0.35
PyYAML==6.0.2
